+++ Content.Shared/CombatMode/Pacification/PacifiedComponent.cs
@@ -1,5 +1,4 @@
 using Content.Shared.Alert;
-using Content.Shared.DeltaV.Implants.Pacifism;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
@@ -15,7 +14,7 @@ namespace Content.Shared.CombatMode.Pacification;
 /// If you want full-pacifism (no combat mode at all), you can simply set <see cref="DisallowAllCombat"/> before adding.
 /// </summary>
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentPause]
-[Access(typeof(PacificationSystem), typeof(PacifismImplantSystem))] // DeltaV add friend PacifismImplantSystem
+[Access(typeof(PacificationSystem))] // DeltaV add friend PacifismImplantSystem
 public sealed partial class PacifiedComponent : Component
 {
     [DataField]
