+++ Resources/Prototypes/Entities/Objects/Misc/subdermal_implants.yml
@@ -342,9 +342,11 @@
   categories: [ HideSpawnMenu ]
   components:
   - type: SubdermalImplant
-  - type: PacifismImplant
-    disallowDisarm: true
-    disallowAllCombat: true
+  - type: AddComponentsImplant
+    componentsToAdd:
+    - type: Pacified
+      disallowDisarm: true
+      disallowAllCombat: true
   - type: AddFactionsImplant
     factions:
     - Hostage
